+++ CHANGELOG.md
@@ -33,6 +33,7 @@ For a full diff see [`fa9c564...master`][fa9c564...master].
 * Extracted `FieldDefinition\Reference` ([#157]), by [@localheinz]
 * Extracted `FieldDefinition\References` ([#159]), by [@localheinz]
 * Extracted `FieldDefinition\Value` ([#160]), by [@localheinz]
+* Extracted `FieldDefinition\Closure` ([#161]), by [@localheinz]
 
 ### Fixed
 
@@ -88,5 +89,6 @@ For a full diff see [`fa9c564...master`][fa9c564...master].
 [#157]: https://github.com/ergebnis/factory-bot/pull/157
 [#159]: https://github.com/ergebnis/factory-bot/pull/159
 [#160]: https://github.com/ergebnis/factory-bot/pull/160
+[#161]: https://github.com/ergebnis/factory-bot/pull/161
 
 [@localheinz]: https://github.com/localheinz